lambdas/requirements-test.txt
black==23.11.0
freezegun==1.2.2
isort==5.13.0
moto==4.1.15
pip-audit==2.6.1
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest==7.4.3
requests_mock-1.11.0
ruff==0.0.284
